+++ .idea/.gitignore
@@ -1,0 +1,3 @@
+# Default ignored files
+/../../Sprint_4./.idea/shelf/
+/../../Sprint_4./.idea/workspace.xml
